pyproject.toml
[project]
name = "compliance-guardian"
version = "0.1.0"
description = "Dynamic compliance guardian for LLM agents"
dependencies = [
    "pydantic>=1.10,<3",
    "rank-bm25>=0.2.2",
    "fastapi>=0.103",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
addopts = "-q"
